presence: Calculate aggregated presence on EVENT_PRESENCE.
Aggregated presence is used to show user status,
so update it to reflect updated status.
Logic to get aggregated presence is almost taken
from web.

Aggregated presence is
* active if user has active status
on any of the client and it's age is less than threshold.
* idle if none client have active status with valid age
and there is a client with status as idle with age less
than threshold.
* else offline.

https://github.com/zulip/zulip/blob/9cc561e442fe441ba8a775e444a1399010cc03ab/static/js/presence.js#L63