flow: Include `types` in our type-checking.
Much of our code that *looks* typed actually hasn't been.  We have a
lot of our types concentrated in this one file `types.js`, and we
haven't marked it as a file Flow should look at -- which means that
when we import a type from there, Flow treats that as a totally
unknown type.  Any property we get off of such a type, or the return
value from any method call on it, then gets the same untyped treatment.

So, mark the file with `@flow`.

Just doing that alone brings 1564 (!) errors.  (Figures based on a
version from the other day; may vary slightly after rebase.)  So in
order to make this mergeable, I dialed it back:
 * out of about 77 types in the file, knocked out about 44 of them
   back down to `any`
 * added 5 more types, all as `any`
 * sprinkled a couple of mysterious `any`s in notifications.ios.js --
   I don't really understand why these give an error while the Android
   counterparts don't.

None of these `any`s represent a regression in our type coverage, I
believe -- they just do the same thing more narrowly that was already
happening as a consequence of the whole file being untyped.  They were
`any`, they're still `any`, but at least some of the other things in
the file now are not.

It'd be a pretty great project for a bunch of people to come change
the various `any`s to real types, keeping the type-checker happy the
whole time.  That'll give us real, truthful types in a lot more of the
code -- which (a) will probably find some live bugs we have right now,
and more importantly (b) will give us much better information about
what the code says as we make changes day to day, which will help us
make fewer bugs in the future.