types: Give a real type for Message, mostly.
This adds coverage for another 333 or so expressions -- over 1% of the
total 28051 in the codebase.  Now at 21648/28051 = 77%.

I say "mostly" a real type because for a couple of properties I left
`any` in place.  Like all the `any` types in the codebase and
especially in `types.js`, these will be good places to follow up and
add real types.

The types in several places were simply wrong.  Note in particular
what `{ message }: Message` means -- it doesn't mean `message` is a
Message. ;-)  Most of these I fixed in place; others are marked with
TODO comments to clean up further.