settings: Rework layout to a more standard style.
Currently, in our various settings screens, we give each item a gray
background and separate items with a wide, 16px, space (in addition to
the 8px of padding within the items around their contents.)  This
looks fairly alien in the context of modern mobile apps, and isn't
particularly attractive either.

Instead, adopt a style that follows Material Design guidelines:
  https://material.io/guidelines/patterns/settings.html
and is also quite similar to settings in iOS and modern iOS apps.
Items have the usual background (white, in light mode), and are
separated by a low-opacity 1px line.

As part of doing this, convert OptionDivider from a function component
to a class component in order to access styles via context.

As is, this causes some settings screens to look a little off; there
are too many dividers, and a few places need dividers that don't have
them, so some followup commits are required.  But this is a net
improvement already.

[greg: Rewrote commit message; tweaked small bits of code style.]