StreamScreen: Modify how padding is applied.
The screen previously used padding prop of Screen component to apply
padding to everything inside, which prevented the OptionDivider from
reaching the full width of the screen.  Now, components inside that
required padding are wrapped in a View with padding, and padding prop
of Screen is not used.

The `padding` style in theme.js was unused, so we repurpose it to
describe the 16px of padding we need here.  This matches the naming
convention of `marginTop`, `smallMarginTop`, and so on; the old value
would have been `smallPadding` by that convention.