refactor: Make selection scrolling more precise
* add checks for user is found and listRef is defined
* setTimeout is now more predictable

If we are selecting a user, we scroll the avatar list to the end
to ensure the last item just added is visible.

Previously we had to wait for some time until the component updates.
We did wait for 300msecs (magic number). Ensuring that the component
already did update we use setTimeout only to allow the UI to refresh
but we no longer need to wait 300msecs.