webview: More conservative start/end message logic
Currently we do an 'union' type of calculation between two points
in time - position before scroll, position after the scroll.

Previously, if we can't find a message node for start we assumed
ID of 0, and for end we assumed ID of MAX_SAFE_INTEGER.

This produces way too many cases where we might want to mark as read
messages from 0 to MAX_SAFE_INTEGER (all visible).

Instead, now we return the opposite values, thus in cases where we
do not find start/end messages we mark as read more strictly.