docs: Document how to make an iOS release.
This comes from the steps I did today and yesterday to get a release
built and uploaded to Apple.

It doesn't cover the laborious debugging of mysterious build failures,
mostly related to Sentry and made worse by Xcode's inscrutable
configuration and painfully slow builds; that story belongs someplace
else, as hopefully it's not a recurring part of the process.