welcome: Style buttons as primary.
This makes them somewhat more prominent -- which is appropriate
because they're the only thing on the page, and the only place we want
the user's attention when on this screen.

I tried making only one button primary and the other secondary, which
in principle might be even better; but with our current button styles,
and with no other context on the screen, it ended up just feeling like
something was broken or one of the buttons was disabled or something.