Process REHYDRATE in Flags Reducer
This fixes a positioning issue.

On startup we did fail to consider the flags of the cached messages.
This was not a problem in cases we invalidate them, but is a problem
when they remain relevant.

We use the same processing code from the MESSAGE_FETCH_COMPLETE to
produce the initial state.