webview: Escape *everything* that isn't intended as HTML.
If a message topic contained HTML metacharacters such as `<` and `&`,
we would nevertheless pass it right through into the rendered HTML.
Because topics are arbitrary text controlled by other users, that is
a bug and potentially a security vulnerability.

I tested with a topic of `hello <strong>topic</strong>`, as a proof
of concept -- and indeed I got the words "hello topic", with the
latter word in bold.

Fix that.  Also fix the many other places our code would recklessly
interpolate into HTML a value which is intended as text, not HTML.

Some of the places we fix happen to never contain HTML metacharacters;
others it'd take some work to be sure, and some look quite suspicious.
Rather than try to be clever to save a few microseconds where the
escaping happens not to matter, apply the only sane policy: *every*
value that isn't intended to function as live HTML, we escape.

The syntax used in this commit is kind of ugly, but it's a darn sight
better than suffering HTML injection, and it was quick to sweep
through systematically.  Nicer syntax to come.