webview: Escape narrow data the same as everything else.
This is a pure refactor -- we do exactly the same escaping, but move
it to right where we interpolate into an HTML string.  This gives it
the same structure we use for all the other escaping, which we chose
in order to make it easier to see what we escape and what we don't.

Remove the now-disused `stringifyNarrow` helper.