compose: Don't overshoot in compose menu open/close animation.
As the compose menu opens, the compose box gets narrower, which forces
it to grow taller.  With overshoot, this would sometimes cause it to
grow a line taller than it eventually needs to, before snapping back.
This caused a jerky effect in the elements above it that have to move
out of its way.

Fixes #2572.