docs: Add that Node 8.x is aka Node LTS.
We recommend this version because it's the latest LTS, but the
download page we link to doesn't discuss what version is the LTS, and
is written in terms of version numbers instead.  Keep the number so it
makes sense of the instructions on the download page, but add a word
about why.

Presumably at some future date a later major version will be the new
LTS; we can update these instructions then.