outbox: Use `local_message_id` rather than `localMessageId`.
Since we are now copying the server event object for EVENT_NEW_MESSAGE
actions, we already have the `local_message_id` property and we do not
need to create a new one called `localMessageId`.

Also rename the corresponding property in two other actions that we
handle together with this one.