language-settings: Decrease check icon size.
Decrease from 24 to 16 logical pixels. Prior to this commit,
the height of the check icon was greater than the height of the
text in LanguagePickerItem; this resulted in the row for the
selected language being taller than the other rows.

We imitate an iOS style found in some settings screens, for
example the upper-right hand screenshot of this Human Interface
Guidelines page:
  https://developer.apple.com/design/human-interface-guidelines/ios/views/tables/

Fixes #2610.