flow: Fix other side of Dispatch type.
The action dispatched doesn't need to be one of the types we've
listed in the Action union; it just needs to be an object with a
string `type`, a constraint already expressed in Redux's `Dispatch`
type.  Let Flow infer what the action type is.

This lets us remove a fixme; the issue was in fact that
react-navigation may dispatch some actions of its own here that
aren't in our `Action`.