notif: Fix incorrect use of last action-creator type.
The use of SaveTokenPushActionCreator was misleading in a couple of
ways:

 * None of the functions that had this type as a parameter actually
   cared that the argument they got would return anything, either a
   SaveTokenPushAction or anything else; they discard the return value.

 * The parameter names were in the wrong order!  Note the swapped
   order at the one caller of these functions, in `realmActions`.  The
   code ends up being right, but only because it doesn't match the
   type.

 * Maybe biggest of all, and related to the first point: the type,
   particularly with its name saying "action creator", sounds like
   this parameter is supposed to be a function that makes an action --
   that we'd then have to do something with after calling it.  That
   makes it concerning when looking at the actual code, which discards
   the return value -- how does this do anything?  In fact, the
   function is called for its side effects, despite the type's name.

Instead, write down a type that matches what this parameter really does.

Also, experiment with a style of putting the type definition locally
in the part of the code that it's concerned with, rather than a large
central "types" module.  This isn't something that a lot of places
across the app need to know or care about.