screen: Add optional `scrollableContent` prop with default value true.
This prop is responsible for deciding whether content wrapper will be
`ScrollView` or `View`.  The Screen component is a very generic
component, it should be customizable for different requirements.
There are some screens in the app in which we don't want full-screen
scroll, like `GroupScreen` where the user chooses recipients for a PM
conversation.  Also `ScrollView` inside `ScrollView` (or `FlatList`,
which is also a kind of scrollView) has performance issues on Android.