timing: Remove group timers.
startGroup() and endGroup() are not used, thus, remove them. In fact,
the startGroup timing logic had a bug: Whenever it is called, it sets
updates the timingMap, but it should only update the timingMap on the
first call for a group. This likely led to incorrect debugging output
at the time when the function was used.