nav: Call fewer selectors on rehydration.
Rehydration reducers should generally avoid calling selectors with the
`payload` of the REHYDRATE action they get.  This is because it can
cause trouble during the initial reyhdration, when the stored state is
empty and this payload will be empty.  The selectors might then try to
access parts of this empty "state".

In our codebase, `navReducers` is the only place where the REHYDRATE
reducer calls a selector, `getInitialNavState`.  `getInitialNavState`
in turn calls the selectors getNav, getAccounts, getAuth, and
getUsersById.  Currently, it is not a problem that these selectors are
called during rehydration: as it happens, they never access the state
deeper than the first layer (`state.foo`), returning undefined, which
can be handled by the code.  Accessing the second layer
(`state.foo.bar`), however, would throw an error.

This will be a problem in an upcoming commit, where getUserById will
call selectors that access the second layer. To prevent the app from
erroring, pull directly into the reducer the logic that handles when
the stored state is empty.