popup: Remove extraneous `maxHeight` prop on a `View`.
This is not a prop that the `View` component supports.  It has never
had any effect, and the Flow 0.75 we'll pull in with RN 0.56 will
register it as an error.

This prop was added in commit ce0395cff when changing the component
used here to a `ScrollView`, and 62ddf77a7 left it around when
changing back to a `View`.