docs: Update all references to `npm` in favor of `yarn`.
We've been relying on Yarn for quite a while now, and in particular
pointing everyone at it in the setup instructions.  Replace all legacy
references to the `npm` command to refer to `yarn` instead.  While
we're at it, take advantage of `yarn`'s convenience feature where
simply `yarn COMMAND` will run a `package.json` script named COMMAND.

A couple of references were really about NPM, the service.  One other
reference is to a script name, `lint`, which we don't actually have;
the script was renamed to `test:lint` in 581aaa7ff back in 2016.