gitignore: Clean out no-longer-relevant CocoaPods entry.
This was added last year in 4096e712d, which also removed these files
from being tracked in the repo.  Shortly after that, in 69613a2a4, we
went on to remove the build code that would create these files; so
nothing creates them anymore, and we don't need the gitignore entry.