gitignore: Add some comments and organization.
This has grown, and it's good to have a bit more info on what
everything is and why it's there.

In the Fastlane section, update the link because the old link is now
just a page pointing to this new one; and also add a line to make our
config match those upstream recommendations exactly.