api: Make `getMessages` return full response
We use often use response processing functions to extract the
meaningful data out of a response. Previously we only cared about
the `messages` property.

Since the adding of `found_newest` and `found_oldest` fields we
now care about the whole response.

This reworks the `getMessages` to return the whole response and
any call to the function to use the `messages` property correctly.