msg actions: Refactor signature of executeActionSheetAction.
The `header` and `title` values are really quite different from the
properties left in `props` -- they identify which action the user
instructed us to take, while the rest are parameters to that action.
So, pull them outside the blob to better reflect that structure.

Also say `isHeader` to make the meaning of that boolean more apparent,
and put it before `title` because it identifies the namespace that
`title` then lives in.

As a bonus and a further sign that this better reflects the structure
of the code, it allows us to eliminate the single-use type
`ExecuteActionSheetParams`, in favor of reusing the same type
`ActionParams` which gets passed to the handlers themselves.