android notif: Change the data structure for saving notification data.
Keep the content of all messages, not only the latest in each
conversation; and keep the message IDs too.

This will give us the data we need in order to support removing
a notification when the server tells us the message has been read.

This is a pure refactor, with no visible effect on behavior; we don't
yet use the additional data.