msg: Clarify what's really of type `Recipient`.
It turns out the callers of `normalizeRecipientSansMe` and
`normalizeRecipient` actually call them with a variety of types;
we've been getting this past the type-checker only because
`Recipient` has been a euphemism for `any`.

What the actual types at the callers have in common is the `email`
property.  So, write a type that means "an array of any subtype of
`{email: string}`".

This is a little tricky because, for good reasons, "any subtype of
(array of `{email: string}`)" is more specific than that; the reason
the type "array of any subtype" works is only because we don't
mutate the array.  So to make the type-checker see why the code is
OK, we have to point that out.

Which Flow happily has a way to do... that is completely absent from
the documentation.  This issue is a little bit like having
documentation, though: https://github.com/facebook/flow/issues/2811

After clarifying these types, we can uncomment the real `Recipient`
type!  I checked just now, and this is in fact the type I saw in
`display_recipients` in the server `message` event for a PM.