android: Bump minimum supported version to 4.4 KitKat (API 19).
Useful information on Android versions can be found here:
https://en.wikipedia.org/wiki/Android_version_history

React Native allows us to support Android 4.1, aka API 16,
aka Jelly Bean, and newer.

We were claiming in the app's manifest that we support 4.1+.  But
our documentation has said 4.4+ for a long time; we've rarely tested
the app on lower versions like 4.1; and some quick testing finds
it's pretty broken there.  And Google Play reports that only a tiny
fraction of our users, 0.5%, are on versions <4.4.

Now, making this official, we set the minimum API level to 19.
Once published to the Play Store, this will ensure only users on
supported versions can download the app.

[greg: added results of testing on 4.1, and user count.]