api: Use `encodeAsUriNoEmptyValues` for API calls
Fixes #2889

This ensures we do not post empty values to API calls like:
```
{
  key: undefined,
}
```

`subscriptionAdd` and `subscriptionRemove` did pass `undefined`
for `principals` when doing a subscription to a stream on the current
user (vs adding or removing other users). This was working before
but is now failing.

Not passing the values at all is the more correct way of doing the
call and also fixes the subscribe/unsubscribe bug.