android notifs: Use different icon for debug build.
Before both debug and release build were having same icon, and both
builds can coexist at the same time in the phone.  So it was hard to
figure out from which build a given notification was generated.

For now, just reuse the debug build's launcher icon as its
notification icon too.  As this icon doesn't follow the normal pattern
for a notification icon, it can look quite odd; but it saves us the
work of making a new icon.  Only contributors will be using the debug
build, so this is fine.

Fixes #2863.