webview: Do not open YouTube and Vimeo links in Lightbox.
Fixes #2895.

We do not support video in the Lightbox screen, nor do we intend to
(no benefit).  Instead treat the links as normal ones if they are
to a video on YouTube or Vimeo.  This matches the behavior of many
other mobile apps.

These are the two types of video links the webapp embeds in messages,
as seen in zulip/zulip's static/js/lightbox.js .  This logic is kind
of fragile, though; for example, we still do the wrong thing on images
inside an embedded tweet.