outbox: Recover from Markdown parse failure and send message anyway.
Fixes #2903.

When sending a message we try to parse the markdown so we can
show an HTML preview.  This is done via this project:
  https://github.com/zulip/zulip-markdown-parser

Sometimes parsing fails, throwing an exception.  In that case
we just don't send the message, which makes no sense.

Instead, wrap the parsing inside a try-catch block.  If the
parsing fails, use the raw message as preview and still send
the message.