api: Clean up parameter-encoding logic.
The `removeEmptyValues` helper mutates its argument, and as a result
so do most of the `api*` functions.  This behavior is hinted at in the
name `removeEmptyValues`, but is quite a surprise for `api*`; so stop
doing that.

Also fold a couple of single-use helpers together, and give the
function here a more specific name.  There are lots of ways we might
(and do) encode different things as URLs, but this one's job is to
encode some parameters to something we could use as the query-part of
a URL, in a way that mimics how it's done for form data on the web:
  https://url.spec.whatwg.org/#concept-urlencoded-serializer
  https://www.w3.org/TR/html52/sec-forms.html#mutate-action-url