compose: Save draft immediately on edit (fix #2861).
Fixes #2861.  On an APP_REFRESH, we were losing the draft the user had
typed into the compose box, which could be pretty frustrating.  The
cause was that we hadn't saved it to our Redux state, in `drafts`,
because we were waiting to do so until the last moment and apparently
missed a case for what "the last moment" could mean.

Instead, save the draft pre-emptively on each edit the user makes.
This fixes #2861, and likely also some cases where the app gets
stopped abruptly but we had plenty of time before that after the user
entered some text.

Also use the new `updateDraft` to replace `tryUpdateDraft`,
significantly simplifying the code.