caught-up: Use `found_newest` and `found_oldest` if available.
Fixes #3338.

Previously we had to use custom caught-up logic to determine if we
have loaded the oldest or newest messages in a narrow.

Back-end now gives us that data while older versions do not.

Keep the old logic when the `found_newest` and `found_oldest`
values are not provided, but use them directly if present.