webview render: Use a proper type for `renderedMessages` prop.
An hour ago this exposed a whole bunch of type errors; we've fixed the
bulk of them in the series of commits leading up to this one.

One error this change exposes is known and about to be fixed; that and
another that seems a bit tricky to fix (and also self-contained), I'm
leaving it in place for now by marking with $FlowFixMe.  That at least
makes them more explicit, which is a small improvement even for these.

One other Flow error, on the use of `match_content` in
`renderMessagesAsHtml`, is fixed with a slightly subtle reassurance
explained in the comment below.