webview render: Give `renderMessages` return value a real type.
The `renderMessages` function returns a moderately complex data
structure, which we give the name `RenderedSectionDescriptor[]`.
But then that identifier is defined as just `any`.  Fix that.

These types are just descriptive, from reading the existing code
and responding to errors from the type-checker; there's room for
some useful refactoring to simplify them, but best not to mix that
with writing down the existing facts.