webview render: Acknowledge Outbox type in parameters.
Also the empty `item` object which `renderMessages` always puts in the
first element of its return value.

This change involves a small edit to actual runtime logic: we adjust
the test in `messageHeaderAsHtml` for "is `item` empty" to something
Flow understands well enough to see that `item` after this point is
one of the non-empty alternatives.