messages: Fix and re-enable most chat reducer tests.
The recent "messages: Normalize `narrows` state`" requires a bunch of
tests to be updated that contain fragments of our message state.  To
keep that commit's diff readable, we just disabled them there; we need
to fix and re-enable them.

This fixes most of the tests in chatReducers-test.js.

Some of the `chatReducers` tests don't apply anymore to chat reducers,
but rather to the new messages reducers introduced in the "Normalize"
commit. Skip those tests for now. The following commit will move them
into a new file for `messageReducers` tests.