flow: Ignore an error with react-navigation on RN 0.56 upgrade.
The RN 0.56 upgrade, with newer versions of Flow and React, produces a
rather complicated Flow issue (Cannot instantiate React.Element ...)
that will likely be fixed with better `flow-typed` declarations in
near future.

Way too cryptic to fix, and it is not pointing to any actual issue
with our code. Ignore for now.