styles: Flatten style props where trivial.
When we need to pass several styles to a component we do it via
an array of style objects. This is not optimal, as this creates a
new array each time. Usually it is not a concern but we should not
overuse it.

There were several styles that, likely because of reworks, were
just an array of a single style. Get them out of the array!