drafts: Merge draftAdd and draftRemove into one action.
Suggested by @jainkuniya and approved by @gnprice.

We don't need two actions to perform the draft updating. Replace
the 'add' and 'remove' actions with a single 'update' action.

The action creator is now very simple and does not need tests. So
remove the 'draftsActions-test.js' file completely.

The logic is exactly the same as before, but is merged in a single
action. The tests changed but test the very same use-cases.