autocomplete: Decide consistently whether to show popup (fix #2791).
We've been deciding whether the popup should be visible using
two criteria:
 (a) There's a prefix character: `:`, `#`, or `@`.
 (b) There's a nonempty filter string following the prefix.

But they can disagree, in both directions (though the possibility of
"!a && b" is really a bug in our `getAutocompleteFilter`).  And
because we apply (b) only through controlling an animation, when we
switch from "!a && b" to "a && !b" the popup suddenly appears, then
disappears gradually over 300ms as prescribed by our animation.

Instead, make one combined boolean "a && b", and use that consistently.

Thanks to @jainkuniya for contributions to debugging this.