webview updates: Revert shallow comparison, return to lodash.isEqual.
This reverts commit e126bacd5.  @armaanahluwalia noticed that after
that commit, we would re-render messages in the webview on lots of
irrelevant changes -- notably on each character the user typed into
the compose box!  Its only purpose in the first place was
optimization; so just revert it, at least for now.

@borisyankov points out that in general, the optimization of doing
only a shallow equality check is only efficient when *some* layer in
the stack takes responsibility for noticing when there's no
substantive change and preventing a lot of wasted work.

React itself does this by diffing the virtual DOM, preventing wasted
updates to the native components (or to the DOM on the web), which
would be the most expensive step.  Our reducers usually also do this
for their respective pieces of the Redux state, which can save a lot
of work in the steps leading up to that.  When we have "derived"
selectors that do more than just pick out a piece of the state, we use
Reselect, whose caching accomplishes the same thing again.

But this is a spot where (a) a change won't be going through React,
but instead straight into the WebView; and (b) these three props come
from derived selectors where we've defeated Reselect's caching
mechanism; see #3015.  As a result, there's no other layer taking
responsibility for this, which is why this code needs to do so.