SpinningProgress: Replace with PNG-based implementation.
React Native's ART library is important and we probably will use
it more in the future. It is the only good option when we want to
achieve some visual elements / styles that on the web we would use
CSS or SVG for instead.

Unfortunately, there is a serious bug in the library on Android,
that has existed for a while.  It becomes 'active' from API level 25
forward, in terms of the app's `targetSdkVersion`.  We want to move
to 26 soon, so this is a blocker for that.

Visually, the issue manifests as black spinning square instead of
the usual progress spinner.

Bug reported to React Native here:
  https://github.com/facebook/react-native/issues/17565

A possible solution looks something like:
  https://github.com/macdoum1/react-native/commit/587848168a80a4cf2244f7c7f6aaff397c675c58

Possible alternatives evaluated:

 * Styles (ala CSS) - previous implementation used that, but was not
   reliable between versions (due to RN issues).

 * Use third-party component - a bit too complicated, and they suffer
   from the same issues. For example:
     https://github.com/bartgryszko/react-native-circular-progress/pull/105

 * Use static images - that is the chosen workaround.

The exact look was recreated in Sketch and exported as PNG. We use
two versions - one with 'brand' color and one 'black'.

Once the bug in ART is fixed upstream, we will revert this change.