react: Begin to reduce over-use of ChildrenArray; prefer React$Node.
Most of the places we use `ChildrenArray` are for props we're just
going to stick straight back through JSX into the `children` prop of a
child component (most often a `View`), ultimately for React to render.

The type of things React can render is `React.Node`, or `React$Node`.
So, when that's what we mean, start using that.  Specifically, do so
in a couple of components where the RN 0.56 / Flow 0.75 upgrade starts
causing type errors if we don't.

In the case of ComponentList, this actually exposes a latent bug: the
type we'd written was compatible with passing it a single component as
`children`, but it implicitly assumes it's really an array.  (Despite
the name, `ChildrenArray<+T>` has a case of just `T`, so
`ChildrenArray<*>` needn't be any kind of array at all.)