message: Ensure `flags` data on incoming new message.
Fixes #3046.

The server returns a key `flags` in the root of the event object
when a new message arrives; and it does not exist on the `message`
object. That is not expected by our code and is likely a thing we
want to change on the back-end for consistency.

The seamless way is for our code to check if the `flags` is on the
server event object and if so move it to the `message` object. This
is done before any processing is done in the `eventMiddleware`.