download: On iOS, authenticate via `api_key` in URL.
Fixes the iOS side of #2618.

The Zulip server generally requires authentication for uploading
and downloading images. Since React Native's CameraRoll doesn't
support setting HTTP Authorization headers at the moment, we pass the
API key as a query parameter. See zulip/zulip@5ddf2614f.

There was previously an `appendAuthToImages` utility method that was
added in cb3198606 and then removed in 9e28ef2e4, and there is
similiar and more robust logic implemented in `appendToAuth` in
src/webview/js.js.

However, it currently isn't easy to share code between the webview
JavaScript and the rest of the app, `downloadFile.js` is used only for
images, and these lines are unlikely to change, so opt to append the
api_key via string concatenation, which is simple and easy.