android: Disable AAPT2 for now, to unblock release builds.
This became the default with Android Gradle Plugin version 3.0.0:
  https://developer.android.com/studio/releases/gradle-plugin#3-0-0
So turning it off returns us to the behavior before we recently
took that upgrade.

Without this line, we run into this issue when attempting to make a
release build:
  https://github.com/facebook/react-native/issues/16906
That issue was closed long ago when an intended fix was merged, but
then that was reverted.  The fix went back in as da6a5e043 upstream,
which is in v0.57.0, so we'll get it when we take that upgrade;
hopefully that then fixes this for real.