download: Tweak URL-with-key logic to be a bit more defensive.
It's not entirely obvious whether it's possible for the `src` URL to
already have a query part, in which case we should be using `&` rather
than `?` as the delimiter.  Rather than make and justify an assumption
that that can't happen, borrow the logic we have for handling this in
`appendAuthToImages` in js.js.

There isn't currently an easy way to share the actual code between
these places (one inside the webview, one in the main RN environment),
so that means copying.  Fortunately this is a tiny amount of code and
its logic isn't likely to need to change, so that isn't so bad.