android notif: Clear notification when all its messages are read.
Implements the Android side of #2634.  iOS still to go.

This isn't yet the ideal behavior -- if you get notified about two
messages and then read one of them, we'd like to update the
notification to reflect only what's still unread.  But it handles the
common case, for a 90% solution: you get some messages that notify
you, then you go catch up on that conversation.

The ideal behavior will wait for a broader overhaul of our Android
notifications code, which we're due for for a number of reasons
already.  The reason that blocks doing the right thing here is that
the data structure we currently keep around (in `conversations`) isn't
nearly adequate for it: when generating the text and metadata of the
notification, in `getNotificationBuilder`, we rely heavily on the
specific GCM message we just received (`getProps()`) for a number of
details including the sender's avatar, the message's topic if
applicable, and the time the message was sent.  After we're finished
handling that message, that data is gone, and we're unable to use it
in future updates.