dependencies: Fix 'yarn' dependency installation
Fixes #3107.

Currently, running `yarn` to install/update dependencies is failing
for some contributors on macOS with Xcode 10.  The cause is `Detox`,
one of the dependencies, fails to build.  Upgrading to the latest
version fixes that problem.

Note on Detox: We are not actively using it but we should, and that is
why I updated the dependency instead of removing it.

They are also finishing up their Android support:
https://github.com/wix/Detox/blob/master/docs/More.AndroidSupportStatus.md