android build: Fix warnings about ignored old build-tools versions.
This eliminates the noisiest of the several kinds of
deprecation-warning spew that our third-party dependencies generate --
one set of lines like this (with varying version numbers) for each of
our 11 dependencies:

  WARNING: The specified Android SDK Build Tools version (25.0.2) is ignored, as it is below the minimum supported version (28.0.3) for Android Gradle Plugin 3.2.1.
  Android SDK Build Tools 28.0.3 will be used.
  To suppress this warning, remove "buildToolsVersion '25.0.2'" from your build.gradle file, as each version of the Android Gradle Plugin now has a default version of the build tools.

A couple of other kinds remain.  But now the complete output of
`./gradlew assembleDebug` can fit on a single screen!