webview: Fix generated JS code if it contains a `\`
In `generate-webview-js` Babel does its job perfectly but then we
put the generated code inside a template literal.  This changes
the `\` symbol's meaning to try and escape the next symbol instead
of being itself; so if the generated code has a `\`, we break.

The next commit edits `js.js` in a way that causes Babel to start
emitting a `\`.  (We add a template literal with an embedded
double-quote, and that gets turned into double-quoted strings where
Babel must escape the embedded double-quote.)  So, fix this bug
before we trigger it.

Before including the generated code, replace any single `\` with a
double `\\` to make sure the escaping results in a single `\`.