UserItem: Mark `avatarUrl` as a maybe-type in the Flow signature.
If user have not set custom (image) avatar, then we are required to
generate gravatar url (hashing email address) on the client side. So in
such cases we don't have avatar url. UserItem takes avatarUrl and
passes it to Avatar, in which it can be null/undefined. So make it
optional in UserItem too.