docs: Start a changelog file.
I already prepare this information to be used in various ways: it goes
on the App Store, the Play Store, and GitHub to accompany each
release, and it's often handy to refer back to.  Start keeping it in a
more convenient in-tree location for all of us to easily refer to.