Fix message loading indicators not showing
When we navigate to a new narrow, there are three cases regarding
the existence of messages:
 * there are no messages
 * there are messages
 * there are currently no messages but fetching

In case (3) we are showing a "message loading" indicator.

Currently though, we are showing it quite rarely. This is caused
by us first navigating to the narrow and then starting to load new
messages (the logic is 'show if no messages but fetching').

We are smart enough to update that state even if the order is
navigation then loading, but it takes some time. Our new handshake
algorithm slows things down a bit too. Thus it seems we no longer
show the loading indicator in many cases.

The solution is simple and effective. First do the fetching, then
do the actual navigation to the narrow.