changelog: Add "highlights" corresponding to notes in app stores.
For release notes on the Play Store and Apple's App Store, the more
detailed list and the references to specific issue and PR numbers
aren't appropriate for the medium.  Include here a copy of the
more-summarized notes I just posted on the Play Store (in our beta
channel) and plan to post on Apple's App Store shortly.

Also note the date of the release.