travis: Don't bother with the cache.
This makes things simpler to reason about.

And, a bit counterintuitively, it isn't appreciably slower, if it is
at all -- took 459s vs 455s, with a sample size of 1 on each side.
Mostly it means we spend 45s to run `yarn`, instead of 45s to download
and unpack the cache.

That makes sense once we read in the Travis docs about the properties
of their caching feature and what it is and isn't good for:

  Large files that are quick to install but slow to download do not
  benefit from caching, as they take as long to download from the
  cache as from the original source.

  https://docs.travis-ci.com/user/caching