travis: Add incantation to fix release build (but don't actually do it.)
With this rather arcane, but harmless, command, the mysterious ENOSPC
error caused in Travis by switching to the release build no longer
happens, and the build completes successfully.

On the downside, the build is about 4 minutes longer -- over 50%
longer -- than with the debug build.  That JS bundle step is *slow* in
Travis.  So although it works, it's not worth the time penalty to do
on every PR; we'll leave it on debug.

We may sometimes want to turn it on in the future, though.  So leave
this here to prevent repeating this debugging session.