message: Add a `DO_NARROW` action, replacing `FETCH_STATE_RESET`.
The old `SWITCH_NARROW` action was refactored in 5f9b13dabd and
the next commit 924d5f6f3, into a standard navigation action
preceded by an action of a new type `FETCH_STATE_RESET`.

I consider the name change to `FETCH_STATE_RESET` to be too rigid, as
it moves us more towards function calls with a single purpose, and
away from state changes that might be handled in several reducers.

Taking the navReducer out of the picture though was a good decision
and even better considering we are decoupling RNav and Redux.

This is almost a pure refactor, except we also add a property
`narrow` to the action.  We will need this in a later commit.