message list: Fix header action sheet so "Unmute topic" ever appears.
The logic in constructHeaderActionButtons which decides whether we
should have an item to mute the topic, or to unmute it, wasn't working
-- it would always show "Mute topic", even if the topic was already
muted.  That's because `mute` was always undefined, because it came
straight from the MessageList component's props... and we never
supplied any data to that `mute` prop.  The root cause of the issue is
that we don't have type-checking on the data flow into those props,
because `connectActionSheet` is in its path and has no type signature.

The case where this logic can actually matter -- where the right
answer can be other than "Mute topic" -- isn't common, because usually
if the topic is muted we don't show its messages in the message list!
But we do in search results; and in fact the bug is easy to reproduce
there.  Fix it.