message list: Remove busted `onReplySelect` callback prop.
The one thing we try to use this callback for is to focus the compose
box after the user choose "Reply" in the message action sheet.
But this doesn't actually work -- the first thing we do is navigate
to the message's narrow, and that causes a new ChatScreen and new
Chat component, which doesn't get the memo.

The UX of the "Reply" action as it is isn't very satisfying, but this
doesn't feel like the right infrastructure for building a better one
anyway; so just rip out this broken bit of plumbing.