flow: Fix "Sketchy null check on boolean which is potentially false."
If we had enabled `@flow strict` it would catch `sketchy-null-bool`
errors. This triggers when you do an existence check on a value that
can be either null/undefined or falsey.

To fix that, make all optional booleans to be non-optional but add
a default prop value of `false` which keeps the behavior but is more
strictly type correct.