apiTypes: Replace an `any` with the much-more-precise `empty`.
This type is... apparently totally undocumented (thanks, Flow), but I
believe it represents "the empty type" in the sense of a type that
contains no values -- so the very bottom of the type hierarchy, and
the most specific possible type.

As a result, if we use that for the indexer property then there is no
possible value that Flow will believe can be used here.