flow: Apply `strict-local` everywhere we can: 306/397 files.
This applies all the strictness rules of `@flow strict`, except
for `nonstrict-import`.  So that lets us avoid adding any new uses
of unsafe patterns in these files, even while they import other
files that use unsafe patterns.

To do this, I first added `strict-local` to every file that has
Flow but not `strict` or `strict-local`:

  sed -i 's/@flow/@flow strict-local/' -- \
      $(git grep -l '@flow [^s]' src/)

Then of course there were a lot of errors.  But the great thing
about this feature is that it doesn't interact between files!
So I just identified the files with errors, and reverted the
changes there:

  git checkout @ -- \
      $(yarn flow --show-all-errors \
        |& perl -lane 'print $1 if (/^Error.* (\S+):\d+:\d+$/)' \
        | sort -u)

This left 306 files gaining this protection, after reverting 91 files.