immutability: Use generics to replace `any` with real types.
This even finds a couple of outright errors in these type signatures,
which were being masked by the `any`s:
 * The `predicate` passed to `filterArray` must take an argument.
 * The `replaceFunc` passed to `replaceItemInArray` must be callable
   with no arguments.

OK; why this file, why now?

I did a topological sort of our modules -- starting with the "leaves"
that don't import anything and are imported by many things, and ending
with ZulipMobile.js which ultimately imports everything.  Here's a
command for that, with the `tools/tsort-modules` script I just added:

  $ tools/tsort-modules list | wc -l
  385

Of those 385 files, this one is #8 on the list:

  $ tools/tsort-modules list | nl -ba | grep immutability
       8  src/utils/immutability.js

And it's the first one that doesn't have `@flow strict` -- indeed, the
first one without `@flow strict-local`:

  $ tools/tsort-modules todo | head -5
  8   src/utils/immutability.js:/* @flow */
  9   src/utils/misc.js:/* @flow */
  14  src/webview/html/template.js:/* @flow */
  15  src/i18n/locale.js:/* @flow */
  16  src/utils/color.js:/* @flow strict-local */

So it's a prime spot to get proper types where they can propagate to
the rest of the codebase, and to enable more use of `@flow strict`.