emoji: Give a specific name to getFilteredEmojiList's parameter.
This function operates on emoji names, so the map keyed on names is
the one it needs.  Make that explicit, and adjust its callers to
match.

Despite changing what selectors the two caller components are
connect'ed to, this is effectively a pure refactor -- the data from
those selectors are used for nothing but to pass to this function, and
this function doesn't actually use the parameter.