emoji: Include realm emoji in autocomplete and add-reaction.
Fixes #2846, aside from the part that restates #2129.

That is: after this fix, you can autocomplete a realm emoji in
composing a message, and you can add a realm-emoji reaction; but that
reaction, like other realm-emoji reactions, still won't show up
properly.  The fix for that part is coming a few commits later.