caughtUp: Correctly handle the counterfactual that `!msg.flags`.
If `msg.flags` is missing entirely, it'd make more sense to treat it
the same as an empty array, rather than an array containing all flags.

This shouldn't actually happen -- as the jsdoc on this `Message#flags`
property in `types.js` explains, this property is actually always
present in a fetch result -- but we have to handle the condition one
way or another to make the types work, and this way is less confusing
for the reader.