progressiveTimeout: Retry sooner at first, then back off farther.
Bring the first nonzero duration down to 0.1s, and the max duration up
to 10s; and move between them exponentially, rather than linearly.

This has us retry sooner when things first start failing, but less in
total -- good if the reason we're retrying is that the local network,
or the server, is overloaded.