login: Rework row spacing in PasswordAuthScreen.
Use `ViewPlaceholder` instead of styles applied to the components.
This is a direction that I find generally helps in RN, considering
the more limited styling options and less immediacy when
implementing the styling.

The benefit is that we do not change the size of the components
and we don't need to support styling for every component we write.
In this case, the next commit benefits from the PasswordInput not
having a margin applied to it.