input: Nail down type of `placeholder`.
We had this as `LocalizableText`... which is actually `any`,
making it a bit worrying to make refactors here because the real
behavior is unpredictable.

Turns out all call sites actually pass strings, so just say that.
When we want the greater abstraction at some call site in the
future, we can sort out what we actually want then and make both
LocalizableText, and the logic in this component that consumes it,
reflect that.