lint: Sort through rules for React.
I was hoping for a `--fix`-able rule for the order of describing
props, state, etc.  Didn't find that (react/sort-comp does part of
the job, but isn't auto-fixable), but record what I learned.

Including that we can drop for free the disabling of
`default-props-match-prop-types`; we have nothing that trips it.