types: Mark types.js as strict-local.
The last real obstacle was LocalizableText; with that fixed, do it!

Kind of puzzling that turning this on in this file causes errors to
newly show up in *another* file... but OK, fix those.  (Flow pointed
out that these checks would be false if the value just happened to be
empty, which indeed isn't what we meant.)