InputWithClearButton: Pass all props to Input again, but explicitly.
Similar story to 960d9ae67 for PasswordInput: we'd recently (in
33f532841) replaced a `{...this.props}` in the `<Input ... />`
with explicitly mentioning the two props that the type annotation
allowed for this component... but in fact the one call site of
this component passed a number of other props as well!  When they
were no longer passed through to `Input`, that caused the search
box to look quite odd and its interaction to not feel right.

Like 960d9ae67, fix this two ways: pass all props through,
and say that in the types and jsdoc rather than asserting
the component takes just two of the props it actually takes.
Follow the subsequent commit 115655cd1 in disallowing the props
we override, too.

Fix a type error this turns up (because the new props type
is exact): we were passing a `textInputRef` that gets
totally ignored.

Also adjust our use of `onChangeText` to allow it to be
optional; that's how it is in the underlying component, it
seems like clearly the preferable API, and it's just as easy
to fix it as to munge the type to make it idiosyncratically
required here.