types: Mark all Props and State types exact, except for errors.
There's never a good reason, in any situation I've seen signs of, for
the Props or State type of a React component to be anything other than
an exact object type.  Where it isn't and making it so gives an error,
that's consistently been a sign of something wrong in our code.

Even in complex situations, where it might be something generic,
parametrized over a type variable (as in a higher-order component
like `withGetText`), or something written in terms of an underlying
component's types (as in `Input`, wrapping upstream's `TextInput`) --
it's always been most appropriate for those to end up being exact
object types too.

So, let's make these types exact.

The edit here was done with:

  perl -i -0pe \
    's/^((?:export )?type (?:Props|State) = \{)(\n.*?\n)?(\};)/$1|$2|$3/msg' \
    src/**/*.js

Then `yarn flow` found a number of errors.  Fixing those is where the
real payoff from exact types will be; but for a start, let's lock
down the majority that didn't find errors.  I used `git checkout @`
on the changes that found errors, to undo those changes; that was
22 files, and this commit is the rest.

We'll fix the majority of those 22 others in the following series
of commits.