appOrientation: Write specific type, not `string`.
Found by Flow upon making a bunch of action types exact.  (Coming
soon; other errors to fix before committing that.)  The object
passed to `dispatch` in `appOrientation` didn't qualify as an
AppOrientationAction, because `orientation: string` is too broad,
and that caused an error once `dispatch` was able to be more picky
about what's an action type.