eventActions: Mark some more array params read-only, etc.
This is certainly our intent; and it prevents Flow from worrying about
unsound use of mutability, once we tell Flow enough about these action
types that it can know enough to get worried.

Background explanation: Flow needs to know that there aren't two
pieces of code X and Y and a shared mutable value V where (a) X might
mutate V, (b) Y might read V, and (c) the type X sees V as is not a
subtype of (nor the same as) how Y sees V.

When X is a function that Y is calling, and V is an argument Y is
passing in, that's what `$ReadOnlyArray<...>` etc. are for -- they
guarantee X won't mutate V, so (a) doesn't apply and we're safe.
That's what we do in 2 of the 3 spots in this commit.

But here we also have the *return value* of `responseToActions`
playing the role of V -- the function is Y, and its *caller* is X.
Flow worries that the returned data might also be visible somewhere
else that's depending on the data staying specifically `EventAction`,
and the caller could mutate it to be other kinds of `Action`.
Here I'm not so sure how to write down the immutability guarantee,
so instead just call these values `Action`.