flags: Revert `initialState` change; make type more specific.
In commit b82ef4d0d "flags: Fix bogus `initialState`" I completely
misread the FlagsState type; as the bit of test data there shows, I
thought it was saying it was a map from message IDs to maps from flag
names to booleans.  In fact it's the transpose of that, and the
`initialState` value we had was perfectly reasonable.

Revert that change; and make FlagsState explicit about what flags are
expected, to match.  This gives us type-checking for e.g.
`flags.starrred`, though any place we do `flags[foo]` sadly just turns
straight into `any` with current Flow.

Thanks to @jainkuniya for pointing out that commit was broken.