android notif: Convert most GCMPushNotifications methods to static.
The paths for showing a notification on one hand, and handling a user
action on it on the other, don't actually have that much in common and
don't particularly belong on the same class.  Making the methods static
will make it easy to move them around.

It's also pretty simple to make them static now, with all the
facade/helper/etc. stuff found in notifyReact having been reduced to
locals there.