android notif: Simplify onOpened path a bit.
The main thing we do with `props` on this path is get the bundle back
out of it with `asBundle`.  (In fact, that's ultimately the only thing
we do with it -- that's exactly what happens after stowing it away on
`InitialNotificationHolder` -- but that part will require a bit more
refactoring to get at.)  So, just pass the bundle around directly.