android notif: Handle getInitialNotification ourselves.
Just do what it says in upstream docs:
  https://facebook.github.io/react-native/docs/native-modules-android

Kind of a lot of steps in the chain, but all very straightforward.

One thing I'm not 100% sure of is storing the notification data as a
static member.  This is exactly what the wix code does (with a small
layer of obfuscation), so it's not a regression, at least.