dev: Remove Reactotron.
This tool badly broke the First Law of Debuggers:

  A debugger may not break the normal functionality of the program.

I don't understand how it produced the effect it did -- but until
we do, we're not going to have the kind of confidence that it won't
do that again that we'd need before re-enabling it for everyone's
normal dev workflows.

Specifically, the following symptom was introduced by commit
69100d1e6 "Reactivate Reactotron!":

  On opening the app cold, with the Chrome debugger open, it
  gets to showing the main screen and partway through actually
  initializing its data... typically up to MESSAGE_FETCH_COMPLETE
  and INIT_STREAMS... then takes about 25-30s before dispatching
  REALM_INIT and INITIAL_FETCH_COMPLETE.

The problem appears to be 100% reproducible starting at that
commit and up to master; 0% reproducible at its parent and
earlier; and it's back to 0% reproducible after this commit,
which takes Reactotron back out.

The very long delay seen in the console is one unambiguous symptom
that's helpful for diagnosing this issue, but naturally the same
delay has bad effects elsewhere in the app's behavior.  In
particular, it badly confounds testing of other features like
tapping a notification -- we wait to navigate there until after we
dispatch that INITIAL_FETCH_COMPLETE, so it looks like the
notification just isn't working.

There's still a delay like 3-6s between INIT_STREAMS and
REALM_INIT, but that's a lot more like the familiar behavior
of the app.