msglist: Make padding and margin independent of font size.
Change the units used for padding and margins from 'rem' to 'px'.
This makes these UI elements not scale with font size change.

While this would be undesirable from a purely designer's point of view,
and not something we would do for a web app, that is a best practice
that the tested apps follow on mobile.  This leaves more space for
the text when enlarged, considering the limited space.

[greg: separated small changes of actual lengths into an explicit
 preceding commit; cleaned out some noise of unrelated other changes]