accountDetails: Fix broken layout.
We are using `ImageBackground`, which uses styles such as `position:
absolute`. Because it is capable of displaying text (/componenets)
over it. So styles such as `alignItems: center, flex: 1` were
resulting items to be present at vertically center of screen instead
of space below image.

By removing this styles and only keeping `justifyContent: center`,
works as expected.

Fixes #3228.