styles: Remove places we no longer use context.
This is a direct follow-up to the parent commit, clearing out places
where that commit removed all remaining uses of context to get styles.
Making it a separate commit just to break up the mass-edit commands at
a logical point for comprehensibility, and because the intermediate
stage doesn't cause any errors in our tools.

Edits performed by these commands:

  perl -i -0pe '
    next if (/this\.context/);
    s/^\s* context:\ Context; \s*?\n//xms;
    s/^\s* static\ contextTypes .*? \}; \s*?\n//xms;
    s/^import\ type\ \{ [^}]* \b\K Context,\ //xms;
    s/^import\ type\ \{\ Context\ .*?\n//xms;
    ' src/**/*.js

  tools/test prettier --fix