account: Replace empty-string fallback with assert, in `getAuth`.
Each remaining call site of this selector makes an assumption that
there is some active account.
 * Many use the return value to make an API request, which will
   never work if `realm: ''`.
 * Many are in UI components displaying logged-in information.
 * A few are in UI components within the authentication flow;
   these still assume a genuine value for `realm`.

And those categories cover it -- I read through all the call sites,
and those that didn't make this assumption I took care of in the
preceding few commits so that they use something else instead.

So, make that assumption explicit.