android notif: Explicitly use system default notification sound.
At present this code has no effect, and has had no effect for a long
time, because we use `Notification.Builder#setDefaults` to override
the sound we set and to use the system's default sound instead.

We're about to fix that, but we're not ready to actually use the
custom Zulip sound this code refers to, because it's too soft and
subtle for this purpose.

So, for now, make this code choose the same default that `setDefaults`
would do.