api: Fix remaining intersection types; more exact object types.
We've found that intersection types in Flow often have puzzling
behavior, and replacing intersections of object types with spreads
produces much more the behavior we'd expect.

Moreover, we want to move almost all our object types to exact object
types -- and for exact object types, the intersection is almost never
what we want.

Fix the last few straggler intersections in this part of the codebase,
and move them to exact object types while we're here.