msglist: Avoid viewport-relative lengths in layout.
We'd been using the `vh` unit -- 1/100 of the viewport's height --
for saying how tall an image in a message should be.

One consequence of this was that images would end up getting scaled
quite differently on a tall device like a Pixel XL or iPhone X than
on a shorter one.

A nastier consequence was that even on a single device, the images'
heights would often change!  In particular, opening and closing the
keyboard makes the viewport much shorter or taller, so the images
would get shrunken or expanded.  Pretty odd-looking if you see it;
and worse, if there were a bunch of images above you in the thread,
they'd cause your scroll position to get messed up.

Fix that by setting the height in px.  The value of 160px is roughly
what the 30vh worked out to on a 5" phone.  (It's 166px or 154px on
an emulated Pixel 2 at default settings, depending on whether the
compose area is hiding or showing the topic input.)

Fixes #3264.

[greg: rewrote commit message]