actions: Simplify action creator types.
It's true these do return values of these specific action types...
but as far as the caller is concerned, that is a detail behind the
abstraction.  So, simplify the code by not specifying that.

Done with the following commands:

    perl -i -0pe '
        # Replace specific action types with `Action`.
        s{\b     [A-Z]\w+Action     \b}{Action}xg;
        # De-duplicate `Action` imports.
        s{\b Action (, \s* Action)+ \b}{Action}xmsg;
      ' \
      src/**/*Actions.js
    git checkout @ src/nav/navActions.js
    tools/fmt

The action creators in `navActions` are an exception... because in
fact `NavigateAction` is *not* a subtype of what we've defined as
`Action`.  There are some type errors that points to, but fixing those
is for another commit at another time.