event: Simplify subscription-event actions by using server's `op`.
We don't gain anything useful by having our own names for each of
these; it's an extra layer of indirection that just makes things
more work to understand.

An ingredient that's important in making this possible is the recent
fixes to `Action` and related types so that Flow understands this code
-- that gave us consistent type-checking without the need to
constantly repeat the names of specific action types, which would have
made this involve much more boilerplate.

One thing I don't love about this version is the extra indentation
caused by the nested `switch`.  On the other hand: note the
type-checking that the `switch (action.op)` is exhaustive!