event: Make a generic "server event" action type; start with streams.
This is another step in conforming our event actions more closely in
general to the events from the server -- which will simplify the code,
reduce indirections and multiple names to trace through, and allow the
code at the API boundary to focus on the cases where our own data
structures have an important, non-boilerplate reason to differ from
those coming from the server.

One thing I don't love here is the level of indentation in the
reducer, from three nested `switch` statements.  There are a variety
of things we might do about that later.