fetch: Retry fetching messages, indefinitely.
Fixes #3281.

This simple change greatly improves the experience with spotty
network connections.

I have a small concern about the request's retry not being cancelled.
I find the compromise to be worth it.  I am working on a more generic
rework of `tryUntilScuccessful` which will allow it to be cancelled
and will include all of this function's invocations.