autocomplete: Show popup only when the input is focused.
Before this change, if you have an autocomplete popup visible for
the message TextInput and then move focus to the topic TextInput,
the first popup stays visible; so you can even wind up with *two*
autocompletion popups (one for message, one for topic) visible at
the same time.

Solve that the same way we do for TopicAutocomplete.