compose box: Drop bogus check in one textInputRef callback.
If this actually has any effect, then it's a bug: we're acting on some
TextInput ref that no longer exists, or isn't ours.

All code that looks at `this.messageInput` already needs to be
prepared for a value of `null`, because that's what we initialize it
to -- and thanks to types, we know that all such code is in fact so
prepared.

This check was introduced in this commit last year:
  1a6aad2fc fix: Do not throw if messageInput reference is undefined
which describes a crash -- and fixes the actual bug, where some code
was blindly invoking a method on `this.messageInput`, at the same time
as adding this check.