android build: Cut now-meaningless `enableAapt2=false` option.
We upgraded to RN v0.57 a while ago, so following the advice in the
comment here, there should be no need for this anymore.

And in fact since the upgrade to the Android Gradle Plugin a few
commits ago in 1132d6ae8, the warning about it has said it has no
effect anyway:

  WARNING: The following project options are deprecated and have been removed:
  android.enableAapt2
  This property has no effect, AAPT2 is now always used.

So, take it out.