account: Add selector to get current info for a given identity.
We'll use this shortly to get the current ackedPushToken.  It could
also be useful for passing around an Identity and upgrading it to an
Auth, with apiKey, just before we actually need to use the latter.

The pattern here with a Map, and a "key" function we make up for
effectively indexing into it with an object by value, is one it'd
probably be useful to generalize and write down as its own small
library.  We already do something very similar for Narrow; this one
demonstrates (a) using a Map rather than a plain object, and (b)
encapsulating the use of a key.  It'd take just a bit more work to
give it the full Map-like interface (rather than just `get`, as we
effectively do here because that's all we need in this case).