user status: Handle 'EVENT_USER_STATUS_UPDATE' action in reducer.
When a `EVENT_USER_STATUS_UPDATE` action is dispatched, handle it
by updating the relevant user key or add a new key.

Currently the web app does update the `away` value separately from
the `status_text` string and that is the immediate plan for mobile
too. It is possible though that both values are changed at once.

 * if an event contains `away: false` this has the meaning of 'clear
   away overriding'
 * if an event contains `status_text: ''` this means 'clear status text'

There is value in keeping our internal state simpler and consistent
with what the back-end gives us as values. To achieve this we make
sure if we get `away: false` and `status_text: ''` we remove the
corresponding keys.