notif: Properly delay getting `usersById` from Redux.
Handling this data here on AppEventHandlers, and passing it down
through the layers of the notification code, was always pretty messy.

With dcbe2ac86 "notif: Move notifications logic from AppEventHandlers"
in 22.0.107 it became a live bug, as we started taking `usersById`
from `props` when the AppEventHandler is first created and keeping it
around, for use when opening notifications.  This means that when the
user opened a group PM notification after the app was already running,
it would never work -- we'd end up navigating to HOME_NARROW.

Thanks to @borisyankov for spotting this bug, by studying that change
in the code.  It readily reproduces, but I think we've never heard
from someone hitting this issue in the wild.

Fix it, by cleaning up our Redux usage here.  The way to get data from
the state is with `getState()`... and a convenient way to pass access
to that down through layers of code is to just pass `dispatch` (which
we need a lot for other reasons), and have the code that needs to read
state be inside a thunk action.

Specifically here, that means `handleNotification` becomes an action
creator; its callers now `dispatch` the resulting thunk action, and
the whole call stacks above it stop knowing or caring about `usersById`.

The boundary between `notification/index` and `notificationActions`
remains kind of messy and porous.  For now, just move the newly-made
action creator from one to the other.