android notif: Get ReactContext in a more reasonable way.
As noted in the TODO-comment this resolves, the existing pattern (a
holdover from the wix library) is pretty absurd -- I don't have much
confidence in exactly what behavior it might have for when that value
gets initialized, and when it might not be.  So I took a look at what
RN upstream's own code does.

The pattern that's ubiquitous upstream for emitting an event looks
like this:

    getReactApplicationContext().getJSModule(RCTDeviceEventEmitter.class)
            .emit("appStateDidChange", createAppStateEventMap());

That first method `getReactApplicationContext` is inherited from
`ReactContextBaseJavaModule` -- most code that emits events is in a
method on some React module class.

Perhaps most of this code should also become methods on some React
module classes.  In any case, though, this codepath is one we reach
from a service, namely NotificationIntentService.  How do we get to a
React-related object from there?

Well, there is one Service class in RN upstream: HeadlessJsTaskService.
It solves this problem like so:

    protected ReactNativeHost getReactNativeHost() {
      return ((ReactApplication) getApplication()).getReactNativeHost();
    }

So, do something a lot like that.