android notif: Completely stop using the wix library.
At this step, it's still in our build config, but none of our code
refers to it, and nor does the built manifest -- so none of its code
ever runs.

The next commit will purge it completely from our Android build.
This one is separate only in order to explicitly remove from our
manifest each of the items it was putting there.