selectors: Drop spurious caching in getCaughtUpForNarrow.
First step on the journey of #3015.  This is one of many selectors
where we use Reselect's `createSelector` in a way where it has no
useful effect at all.

This case is one where if we did use `createSelector` correctly, the
caching would still have minimal value -- the body of the selector is
scarcely more expensive than the cache logic, likely even cheaper.
So make things simple and just drop it.