selectors: Fix caching on outboxMessagesForNarrow.
Another early step on the journey of #3015.  We've been using
`createSelector` here in such a way that it has no useful effect
at all.

Unlike in the parent commit, the body of the selector does a fair
amount of work, and it sure would be nice to have caching for it.
So rearrange things to put `createSelector` on the outside,
demonstrating the pattern discussed in #3015.

This example looks to be about the most complex one we have, so
the fact the pattern works here is a good sign!

One wrinkle of the pattern which hasn't been discussed on #3015:
the `(state, narrow) => narrow` thing.  This is a bit of a hack,
which is there because although Reselect has this facility for
getting parameters ("props") down to selectors *under* it:
  https://github.com/reduxjs/reselect#accessing-react-props-in-selectors
it doesn't seem to have a reasonable way to get them to the
`createSelector` callback itself.

NB the cache has only one slot, and so it will be missed any time we
switch from one narrow to another.  That seems fine -- we'll almost
always have a single narrow we care about at a time.  In any case
this is certainly an improvement over no working cache at all.