actionTypes: Make some vague types more specific.
For each of the `[string]: any`, I first changed `any` to `mixed`,
which means no assumptions at all about the value are permitted;
then the resulting Flow errors told me exactly what assumptions
our code is currently making about these types.

Those assumptions all looked very plausible; and given that the
current code in these areas works as far as we know, it seems like
they're empirically satisfied.  So I just wrote down those current
assumptions.

There are clever things we can do to give a proper type to
`EventUpdateDisplaySettingsAction`, much like how we currently manage
to type `Action` itself.  Or perhaps more to the point, like how we
type `StreamEvent`.  But those are more work and we don't need that
now, so we'll cross that bridge when we come to it.