Avatar: Do not use `TextAvatar`, and simplify.
`TextAvatar` was used as a fallback if we can not find an avatar
for the user.

That was possibly effective very long time ago but right now we
always use as a fallback the gravatar of a user or the auto-generated
gravatar shape (there is always one!).

Changes in the recent commit ef209dbb8 inadvertently changed the
avatarUrl checks for '' into `avatarUrl === 'string'`, which is
never true.

This makes our use of `TextAvatar` not necessary (and a mess!)

We also don't need the `name` prop, so it is removed from the
component and all components that use `Avatar`.