travis: Set TERM=dumb to avoid Gradle spew.
See discussion of `--console` in the Gradle CLI doc:
  https://docs.gradle.org/current/userguide/command_line_interface.html

With the default settings in Travis, Gradle decides it has a
real terminal, and spits out a bunch of "rich output" that is
indeed very handy when running interactively in a real terminal.
But this involves constantly overwriting a line with an updated
line, and the Travis log doesn't handle that right and instead
shows the various versions of the line one after another,
making a giant mess.  Like this:
  https://travis-ci.org/zulip/zulip-mobile/builds/496248778

By setting the environment variable `TERM` to `dumb`, we
cause Gradle to skip the "rich" output.  Together with `-q`,
this makes it totally silent in the success case.