GroupAvatar: Rename from TextAvatar.
Renames the component from name explaining what it does to a name
that describes its purpose.

While previously used as a fallback for a single user's avatar if
it was missing, we are now using it only for group avatars.

We had plans for a graphic group avatar, consisting of either
several small circles or one with slices of different users' avatrs.
This is a lower priority task but if we do implement it we will not
need to rename this component.