nav: Fix type NavigationRouteState.
This type isn't really ours, but describes some objects given to us by
react-navigation.  We have some defensive code in navSelectors which
caused me to wonder whether the type is really as specific as this.

So I studied the code that it describes: `state.nav` is controlled by
our `nav` reducer, and that basically delegates to `getStateForAction`
in react-navigation/src/routers/StackRouter.js ... of which in turn
the interesting part is `getInitialState`.  And in fact it turns out
that this type is not quite right!  Fix that.