getAvatarFromMessage: Take into account current realm.
Make the function take `realm` as parameter and take it into account
with the `getFullUrl` function, thus making the logic equivalent with
the one in `getAvatarFromUser`.

This fixes potential issues with it not making an URL absolute.

The reason this has not been an issue currently is that the function
is used currently in only two places:
 * `Lightbox` which later calls `getFullUrl` in `UserAvatarWithPresence`
 * `messageAsHtml` which renders in the WebView's HTML and the
   WebView itself has `baseUrl=auth.realm`