android notif: Rename MessageFcmMessage (was PushNotificationsProp).
An instance of this class isn't really a set of "props" in any
meaningful sense, as the parameter and variable names often say;
and it's definitely not a singular "prop", as its own name says.

What it is is a result of parsing an FCM message, in the specific
case where the message has event type `message`, which means it
describes a Zulip message to show a notification for.  Rename
appropriately, and add some javadoc to explain.