android: Check in code-style config for Android Studio.
This will help us use the editor's auto-formatting more
systematically, because we'll be able to adjust it where needed and
have confidence in preserving and sharing the adjustments.

These are the relevant files as they exist in my working tree.
I don't totally understand where the Project.xml came from, or
what all of the settings there mean; but

 * all of them also appear in my "defaults" file at
     ~/.AndroidStudio3.3/config/codestyles/Default.xml

 * all of the ones that I do understand (in particular the
   one for Java, and the CONTINUATION_INDENT_SIZE for XML)
   are settings I know I want.