title-buttons: Drop use of placeholder when no button needed.
It's not clear what purpose this serves.  The hardcoded width of 44px
has stayed constant even as we've adjusted NAVBAR_SIZE from 52px to
58px.

Probably it would be good to have some margin between the title text
and the buttons to the right... but that should apply even when both
buttons are present.