Touchable: Fix a debugging change that got squashed in, oops.
This change *was* a separate one-line commit, with a commit message
starting literally with "NOMERGE", which I had at the start of my
branch while developing the last few commits, in order to more easily
test them.

Somehow I apparently squashed it into the main substantive change
that followed it, which became ec32af1a4.  Oops. :-/

Hmm, in fact: it looks like this happened at the very end when I did a
`git rebase -i` and dropped the NOMERGE commit.  There was a rebase
conflict because the main change had touched an adjacent line...
and `git rerere` offered a recorded resolution of the conflict, from
when I'd previously rebased the NOMERGE commit down to the start of
the branch, having been originally made after the main change...
and I figured that meant it was good, and didn't look too closely
at the recorded resolution.  Lesson: when `git rerere` gives you a
gift, look it closely in the mouth.

Anyway, revert this unintended change.