emoji: Avoid perverse-`reduce` pattern in new selector code.
This is a case where fitting the logic into the shape of `reduce`,
when the logic doesn't actually fit how `reduce` is intended to work,
made things extra complicated.  We really want to iterate over two
objects in turn, but we went and built a whole new object from two
object-spreads, just to have a single thing to iterate over.

Instead, as we iterate over some input and mutate a loop accumulator,
express that naturally with `forEach`.  Now it's perfectly natural to
iterate over two different things in turn, too.