loading: Cut unmaintained `streams` property.
The parent commit removed a place where we set this property,
raising the question of whether it was still maintained at all
and therefore whether it should be removed.

In fact, it turns out this property has never been maintained!
Outside of tests, it's only ever been set to `false` -- never
any other value.  It's never been used, either.

So, cut it out.